Parametrize the cinematic duration
The intro cinematic and subsequent setup events are currently hardcoded with a few integers in various files.

Instead, I define the variable "introSeconds" and pass it to the various setup tasks that require it, replacing hard coded numbers with arithmetic operations so that it can be changed easily.

Tested for durations from 30 seconds to 60.

Note: if introSeconds = 60, basically nothing changes.